feat(admin): Implement all remaining design doc features
New shared components:
- TypeToConfirmDialog (type phrase to confirm high-risk actions)
- useAdminKeyboardShortcuts (global keyboard shortcuts with modifiers)
- useSessionTimeout (activity tracker + 5-min warning before expiry)
- DensityProvider/useDensity (compact/comfortable table density toggle)

New admin features:
- ListingDetailModal (full post detail dialog with moderation actions)
- AdminNotificationsBell (bell icon with unread badge + dropdown)
- Bulk Actions on Listings (multi-select checkboxes + floating bulk bar)
- j/k keyboard navigation on Listings table
- Density toggle button in admin top bar
- Session timeout warning banner

Backend - RBAC:
- PermissionEntity (granular permission keys with categories)
- RolePermissionEntity (role-to-permission junction table)
- AdminPermissionsController (GET all, GET by role, PUT update role permissions)

Backend - Fraud Detection:
- AdminFraudDetectionController (4 rule-based signals: rapid registration,
  duplicate listings, price anomalies, review bombing)

All builds pass: dotnet build (0 errors), tsc --noEmit, vite build.

+++ apps/api/src/Api/Contracts/Responses/PostResponseDTO.cs
@@ -15,7 +15,7 @@ public class PostResponseDTO
     public int Status { get; set; }
     public System.DateTime CreatedAt { get; set; }
     public bool IsDeleted { get; set; }
-    public int? Views { get; set; }
+    public long? Views { get; set; }
     public int? CityId { get; set; }
     public int? AreaId { get; set; }
     public string Location { get; set; } = string.Empty;
